Fixed the placement of Officers
The Officers can only be selected from the Group that is defined. This binds the controls conceptually. Where it was placed the officers table looks out of place but also it took me some while to work out why I could not find the people I needed on the drop down. It left me confused, and exploring until I noticed the the selection was restricted to the Group! Doh! The UI should ideally be more intuitive, and IMHO selecting a group and then the officers from that group is a sensible workflow, making clear why a Group needs to be selected and on what basis Officers are available.

The User selector is currently broken:

https://github.com/tendenci/tendenci/issues/954

And in any case should be a live one that is responsive to the selected Group.

There is  layout issue on this that I have not reco9lved and hence it remains a WIP (Work In Progress).

The layout issues are 2:

1) The subtle: The Officer table does not align on its left boundary with other fields. It looks slightly out of place. Have not been able to divine why, the amount of CSS at play here is not trivial.

2) The Gross. There are three boxes at right of the tendenci2020 theme (Get Updates, Site Search and Recent News). The width of the Officers table is such that on a narrow screen (portrait), it overlaps (underlaps) with these. This is a theme design issue and IMHO if at all possible these three boxes should be floating right as they are but wrap around things in the body that need space. in this case they should slip down below the Officers table I guess, Will take some study of the CSS to find a solution there.

Also implemented a better layout for committee pages

Subjective but I put the committee members after mission before content, and I removed the hone columns (who is publishing a committees phone numbers by default today? Easy to add, to a local template but the default should most certainly not assume phone number exposure.