Made the Members Quick List much more useful
Added important columns: Member number, Membership Type and Groups.
Added click-able headers that sort the table
TODO: The group header is not click-able. Against all documentation for some reason I cannot convince this version of Django (and it could be because it's old) to understand "user__group_set" which works a charm ion the template, but annotate and order_by both barf at it complaining that group_set can't be found as a field dumping a list of available fields (from which it is conspicuously missing indeed)
Fixed the confusion about user_groups
Turns out group_set didn't work, probably because of an ambiguity, namely the Group model has three whole foreign references to User and none hd related_names. So added a related_name for the user_groups and suddenly everything worked.
Implemented reversible sorting.
Could not find a way to do that with a generic user_group list, so only in the postgresql case is the user_group column orderable.
Fixed unused imports
Fixed a bug in the header link test
And ordered group list by id