Made the Members Quick List much more useful
Added important columns: Member number, Membership Type and Groups. 
Added click-able headers that sort the table

TODO: The group header is not click-able. Against all documentation for some reason I cannot convince this version of Django (and it could be because it's old) to understand "user__group_set" which works a charm ion the template, but annotate and order_by both barf at it complaining that group_set can't be found as a field dumping a list of available fields (from which it is conspicuously missing indeed)